Plan trilogy talks for CRC and fix compilation
As from last time I worked on these presentation I changed the
TeXnicalities package introducing there a versioning and a compatibility
option, compilation of these presentations broke as the package was
loaded without the option. It is better to have the version fixed to
avoid this happening again in the future. Also a common command
definition I often use was moved to the package and had to be removed
from these files.